BERT-sentiment-analysis-service/requirements.txt
confluent-kafka
transformers
torch
PyYAML